fix(deps): require python 3.6 (#46)
`google-cloud-access-context-manager` is a proto only package that is only used by `google-cloud-asset`. The last version of `google-cloud-asset` that supports python 2.7 is `1.3.0`. See the [setup.py of version 1.3.0 here](https://github.com/googleapis/python-asset/blob/v1.3.0/setup.py#L34). In version `1.3.0` of `python-asset`, the version of `google-cloud-access-context-manager` that is required has an upper bound of `< 0.2.0dev`. Assuming that the SemVer minor version of `google-cloud-access-context-manager` increases after this PR, this change should not break users that are still on `python-asset` version `1.3.0`.